Generalize substring search relevance + apply to most dropdown searches
This is a generalization of the search improvements made in
  7da2edeaa0f6cf7794f4f8f68960497dc1be893c and
  92235222dba4e5d0c7713d53ef12e0f10e371b83
And has been applied to the dropdown searches for producers, staff, tags
and traits.

For all those searches, exact matches are listed first, followed by
prefix matches, and then substring matches. Relevance is currently only
based on the primary name/title and ignores aliases (except for staff).
This is fixable, but not trivial, and I'm not sure it's all that useful.